Added wishlist, created page for displaying OfferBundles(+other functionalities for them), added image for Carousel1(for a specific bundle)+ some small changes in other jsps

+++ src/main/java/com/nextgenpos/demo/common/OfferBundleDto.java
@@ -7,20 +7,20 @@ import java.util.Collection;
 
 public class OfferBundleDto {
 
-    private Integer id;
+    private Long id;
     private String name;
     private Collection<OfferItem> offerItemList;
 
     private CarouselPhoto carouselOfferPhoto;
 
-    public OfferBundleDto(Integer id, String name, Collection<OfferItem> offerItemList, CarouselPhoto carouselOfferPhoto) {
+    public OfferBundleDto(Long id, String name, Collection<OfferItem> offerItemList, CarouselPhoto carouselOfferPhoto) {
         this.id = id;
         this.name = name;
         this.offerItemList = offerItemList;
         this.carouselOfferPhoto = carouselOfferPhoto;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
